Updated App to render Map as a component imported from another file
